pyproject.toml
[tool.poetry]
name = "image_pipeline"
version = "0.1.0"
description = ""
authors = ["alexgQQ <alex.grand@beatport.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
opencv-python = "^4.5.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
